requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
colorlog==4.0.2
cryptography==2.6.1
cycler==0.10.0
decorator==4.4.0
Flask==1.0.2
Flask-SQLAlchemy==2.4.0
get==2019.4.13
idna==2.8
ImageHash==4.0
imageio==2.5.0
imutils==0.5.2
itsdangerous==1.1.0
Jinja2==2.10.1
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.0.3
#mysql==0.0.2
mysql-connector-python==8.0.16
#mysqlclient==1.4.2.post1
networkx==2.3
numpy==1.16.3
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==6.0.0
post==2019.4.13
protobuf==3.7.1
public==2019.4.13
pycosat==0.6.3
pycparser==2.19
pymediainfo==4.0
pyOpenSSL==19.0.0
pyparsing==2.4.0
PySocks==1.6.8
python-dateutil==2.8.0
python-dotenv==0.10.1
pytz==2019.1
PyWavelets==1.0.3
query-string==2019.4.13
request==2019.4.13
requests==2.21.0
ruamel-yaml==0.15.46
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
SQLAlchemy==1.3.3
tqdm==4.31.1
urllib3==1.24.1
Werkzeug==0.15.2

